fix(*): separate Ingredien component from Ingredient type
