+++ src/components/burger-ingredients/types.ts
@@ -1,3 +1,5 @@
+import { Ingredient } from "../../types/types";
+
 export enum IngredientTypeName {
     BUN = 'bun',
     MAIN = 'main',
@@ -8,21 +10,6 @@ export type IngredientCategory = {
     value: string;
     title: string;
     typeName: IngredientTypeName;
-};
-
-export type Ingredient = {
-    _id: string;
-    name: string;
-    type: string;
-    proteins: number;
-    fat: number;
-    carbohydrates: number;
-    calories: number;
-    price: number;
-    image: string;
-    image_mobile: string;
-    image_large: string;
-    __v: number;
 };
 
 export type IngredientWithCounter = Ingredient & {
