+++ src/components/ingredient-card/index.ts
@@ -1,0 +1,1 @@
+export { IngredientCard } from './ingredient-card';
