+++ src/components/ingredient-card/ingredient-card.tsx
@@ -1,11 +1,11 @@
-import { IngredientProps } from './types';
 import styles from './style.module.css';
 import {
     Counter,
     CurrencyIcon,
 } from '@ya.praktikum/react-developer-burger-ui-components';
+import { IngredientCardProps } from './types';
 
-export const Ingredient = ({ data }: IngredientProps) => {
+export const IngredientCard = ({ data }: IngredientCardProps) => {
     const contentClass: string = `ml-4 mt-6 ${styles.content}`;
     const imageClass: string = `pl-4 pr-4 pb-1 ${styles.image}`;
     const priceClass: string = `pb-1 ${styles.price}`;
@@ -13,11 +13,13 @@ export const Ingredient = ({ data }: IngredientProps) => {
 
     return (
         <div className={contentClass}>
-            {data.quantity && data.quantity > 0 && <Counter
-                count={data.quantity}
-                size="default"
-                extraClass='m-1'
-            />}
+            {data.quantity && data.quantity > 0 && (
+                <Counter
+                    count={data.quantity}
+                    size="default"
+                    extraClass="m-1"
+                />
+            )}
             <img
                 src={data.image}
                 className={imageClass}
