+++ src/components/ingredient-card/style.module.css
