+++ src/components/ingredient-card/types.ts
@@ -1,6 +1,6 @@
 import { IngredientWithCounter } from '../burger-ingredients/types';
 
-export type IngredientProps = {
+export type IngredientCardProps = {
     data: IngredientWithCounter;
     onClick?: () => void;
 };
